+++ sixth-springboot/src/main/java/org/sopt/www/firstspringboot/security/CustomJwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package org.sopt.www.firstspringboot.jwt;
+package org.sopt.www.firstspringboot.security;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
