+++ sixth-springboot/src/main/java/org/sopt/www/firstspringboot/security/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package org.sopt.www.firstspringboot.jwt;
+package org.sopt.www.firstspringboot.security;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -6,6 +6,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import org.sopt.www.firstspringboot.external.jwt.JwtTokenProvider;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
@@ -14,7 +15,7 @@ import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
 
-import static org.sopt.www.firstspringboot.jwt.JwtValidationType.VALID_JWT;
+import static org.sopt.www.firstspringboot.external.jwt.JwtValidationType.VALID_JWT;
 
 @Component
 @RequiredArgsConstructor
